+++ example/views.py
@@ -23,8 +23,7 @@ from example.serializers import (
     EntryDRFSerializers,
     EntrySerializer,
     ProjectSerializer,
-    ProjectTypeSerializer
-)
+    ProjectTypeSerializer)
 
 HTTP_422_UNPROCESSABLE_ENTITY = 422